Set a scope that defines the subset of records in the nested set
Some records in a table should never participate in the
nested set, and should not be considered one of the roots.

First, define a scope on the class to identify records
that should be included in the nested set.
Second, pass the name of the scope as the :subset on
the acts_as_nested_set class method call.